requirements.txt
anthropic>=0.25.0
python-telegram-bot>=20.0
garminconnect>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
apscheduler>=3.10.0
python-docx>=1.1.0
PyPDF2>=3.0.0
numpy>=1.26.0
